Project initial structure created. JSON configuration implemented. Project organized into separate packages.
