+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/tothbence9922/kawe/kawe
+
+go 1.17
